Added the follow button to profile page and added a page to list followers on a personal board
